Eliminate generate subcommand
