+++ Tests/swift-package-list-tests/ExecutionTests.swift
@@ -39,23 +39,4 @@ final class ExecutionTests: XCTestCase {
         XCTAssertNotNil(Int(minorVersionNumber))
         XCTAssertNotNil(Int(patchVersionNumber))
     }
-    
-    func testGenerateSubcommandExecution() throws {
-        let swiftPackageListBinary = productsDirectory.appendingPathComponent("swift-package-list")
-        
-        let process = Process()
-        process.executableURL = swiftPackageListBinary
-        process.arguments = ["generate", "--help"]
-        
-        let pipe = Pipe()
-        process.standardOutput = pipe
-        
-        try process.run()
-        process.waitUntilExit()
-        
-        let data = pipe.fileHandleForReading.readDataToEndOfFile()
-        let output = String(decoding: data, as: UTF8.self).trimmingCharacters(in: .whitespacesAndNewlines)
-        
-        XCTAssertTrue(output.contains("USAGE: swift-package-list generate"))
-    }
 }
